+++ tests/BoscaliSummer.Tests/Program.cs
@@ -6,7 +6,6 @@ using BoscaliSummer.Tests.Features.FireAndDestruction;
 using BoscaliSummer.Tests.Features.Radio;
 using BoscaliSummer.Tests.Features.Progression;
 using BoscaliSummer.Tests.Features.Support;
-using BoscaliSummer.Tests.Features.UrbanCombat;
 using BoscaliSummer.Tests.Framework;
 
 namespace BoscaliSummer.Tests
@@ -17,8 +16,6 @@ namespace BoscaliSummer.Tests
         {
             FrameworkTests.Run();
             ImpactScorchTests.Run();
-            StrongholdTests.Run();
-            TroopDeploymentTests.Run();
             RadioTests.Run();
             ProgressionTests.Run();
             SupportTests.Run();